requirements.txt
numpy

pyserial
# pyvisa
# pyvisa-py